+++ include/iot/data.h
@@ -1074,6 +1074,7 @@ extern iot_data_t * iot_data_from_json (const char * json);
  * the JSON. See also iot_data_to_json.
  *
  * @param  json  Input json string
+ * @parm ordered Whether returned map is ordered by position in json
  * @return       Pointer to data of type iot_data if input string is a json object, NULL otherwise
  */
 extern iot_data_t * iot_data_from_json_with_ordering (const char * json, bool ordered);